Add offer urls as variables for hypervisor terraform plan
* For openstack-hypervisor terraform plans, add variables
for offer urls to be consumed. This is to avoid occurence
of bug https://bugs.launchpad.net/juju/+bug/2085310 if the
remote offer is removed by force
* Remove openstack terraform plan backend config from
openstack hypervisor terraform variables
* Update DeployOpenStackControlPlaneStep to update all
offer urls in terraform tfvars
* Update Telemetry and InstanceRecovery features to pass
necessary offer ruls during enable/disable for step
ReapplyOpenStackHypervisorTerraformStep
* Update Telemetry and InstanceRecovery feature to add step
to remove RemoteSAAS as a wordaround for
https://github.com/juju/terraform-provider-juju/issues/473
* Move RemoveSaasApplicationsStep to sunbeam/steps/juju.py
from observability feature to make it reusable
* Add saas_apps_to_delete parameter to RemoveSaasApplicationsStep
to filter the saas applications that need to be removed
