+++ cloud/etc/deploy-openstack-hypervisor/variables.tf
@@ -53,18 +53,62 @@ variable "machine_model" {
   type        = string
 }
 
-variable "openstack-state-backend" {
-  description = "backend type used for openstack state"
-  type        = string
-  default     = "local"
-}
-
-variable "openstack-state-config" {
-  type = map(any)
-}
-
 variable "endpoint_bindings" {
   description = "Endpoint bindings for openstack-hypervisor"
   type        = set(map(string))
+  default     = null
+}
+
+# Mandatory relation, no defaults
+variable "rabbitmq-offer-url" {
+  description = "Offer URL for openstack rabbitmq"
+  type        = string
+}
+
+# Mandatory relation, no defaults
+variable "keystone-offer-url" {
+  description = "Offer URL for openstack keystone identity-credentials relation"
+  type        = string
+}
+
+variable "cert-distributor-offer-url" {
+  description = "Offer URL for openstack keystone certificate-transfer relation"
+  type        = string
+  default     = null
+}
+
+variable "ca-offer-url" {
+  description = "Offer URL for Certificates"
+  type        = string
+  default     = null
+}
+
+# Mandatory relation, no defaults
+variable "ovn-relay-offer-url" {
+  description = "Offer URL for ovn relay service"
+  type        = string
+}
+
+variable "ceilometer-offer-url" {
+  description = "Offer URL for openstack ceilometer"
+  type        = string
+  default     = null
+}
+
+variable "cinder-ceph-offer-url" {
+  description = "Offer URL for openstack cinder-ceph"
+  type        = string
+  default     = null
+}
+
+# Mandatory relation, no defaults
+variable "nova-offer-url" {
+  description = "Offer URL for openstack nova"
+  type        = string
+}
+
+variable "masakari-offer-url" {
+  description = "Offer URL for openstack masakari"
+  type        = string
   default     = null
 }
