+++ sunbeam-python/sunbeam/features/observability/feature.py
@@ -80,6 +80,7 @@ from sunbeam.features.interface.v1.openstack import (
     OpenStackControlPlaneFeature,
     TerraformPlanLocation,
 )
+from sunbeam.steps.juju import RemoveSaasApplicationsStep
 from sunbeam.steps.k8s import CREDENTIAL_SUFFIX
 from sunbeam.utils import click_option_show_hints, pass_method_obj
 
@@ -224,113 +225,6 @@ class UpdateObservabilityModelConfigStep(BaseStep, JujuStepHelper):
             LOG.exception("Error updating Observability Model config")
             return Result(ResultType.FAILED, str(e))
 
-        return Result(ResultType.COMPLETED)
-
-
-class RemoveSaasApplicationsStep(BaseStep):
-    """Removes SAAS offers from given model.
-
-    This is a workaround around:
-    https://github.com/juju/terraform-provider-juju/issues/473
-
-    For CMR on same controller, offering_model should be sufficient.
-    For CMR across controllers, offering_interfaces are required to
-    determine the SAAS Applications.
-    """
-
-    def __init__(
-        self,
-        jhelper: JujuHelper,
-        model: str,
-        offering_model: str | None = None,
-        offering_interfaces: list | None = None,
-    ):
-        super().__init__(
-            f"Purge SAAS Offers: {model}", f"Purging SAAS Offers from {model}"
-        )
-        self.jhelper = jhelper
-        self.model = model
-        self.offering_model = offering_model
-        self.offering_interfaces = offering_interfaces
-        self._remote_app_to_delete: list[str] = []
-
-    def _get_remote_apps_from_model(
-        self, remote_applications: dict, offering_model: str
-    ) -> list:
-        """Get all remote apps connected to offering model."""
-        remote_apps = []
-        for name, remote_app in remote_applications.items():
-            if not remote_app:
-                continue
-
-            offer = remote_app.get("offer-url")
-            if not offer:
-                continue
-
-            LOG.debug("Processing offer: %s", offer)
-            model_name = offer.split("/", 1)[1].split(".", 1)[0]
-            if model_name == offering_model:
-                remote_apps.append(name)
-
-        return remote_apps
-
-    def _get_remote_apps_from_interfaces(
-        self, remote_applications: dict, offering_interfaces: list
-    ) -> list:
-        """Get all remote apps which has offered given interfaces."""
-        remote_apps = []
-        for name, remote_app in remote_applications.items():
-            if not remote_app:
-                continue
-
-            LOG.debug(f"Processing remote app: {remote_app}")
-            for endpoint in remote_app.get("endpoints", {}):
-                if endpoint.get("interface") in offering_interfaces:
-                    remote_apps.append(name)
-
-        return remote_apps
-
-    def is_skip(self, status: Status | None = None) -> Result:
-        """Determines if the step should be skipped or not."""
-        super().is_skip
-
-        model_status = run_sync(self.jhelper.get_model_status(self.model))
-        remote_applications = model_status.get("remote-applications")
-        if not remote_applications:
-            return Result(ResultType.SKIPPED, "No remote applications found")
-
-        LOG.debug(
-            "Remote applications found: %s", ", ".join(remote_applications.keys())
-        )
-        if self.offering_model:
-            self._remote_app_to_delete.extend(
-                self._get_remote_apps_from_model(
-                    remote_applications, self.offering_model
-                )
-            )
-
-        if self.offering_interfaces:
-            self._remote_app_to_delete.extend(
-                self._get_remote_apps_from_interfaces(
-                    remote_applications, self.offering_interfaces
-                )
-            )
-
-        if len(self._remote_app_to_delete) == 0:
-            return Result(ResultType.SKIPPED, "No remote applications to remove")
-
-        return Result(ResultType.COMPLETED)
-
-    def run(self, status: Status | None = None) -> Result:
-        """Execute to remove SAAS apps."""
-        if not self._remote_app_to_delete:
-            return Result(ResultType.COMPLETED)
-
-        model = run_sync(self.jhelper.get_model(self.model))
-
-        for saas in self._remote_app_to_delete:
-            LOG.debug("Removing remote application %s", saas)
-            run_sync(model.remove_saas(saas))
         return Result(ResultType.COMPLETED)
